requirements.txt
requests==2.31.0
python-dotenv==1.0.0 
fastapi==0.115.12
uvicorn==0.34.0
redis==5.2.1
langchain==0.3.22 
langchain-community==0.3.20
langchain-google-genai==2.1.2
pydantic>=2.5.2
python-multipart>=0.0.6
unstructured
pdfminer.six
